feat : 역을 추가할 시, 이미 존재하는 역 이름이 있는지 검증한 후 추가하도록 로직을 수정하고 역 이름으로 역 객체를 찾는 로직을 구현한다.
- findStation() : Optional 객체를 반환
